# Improve change tracking performance