@Turnerj 2023-10-01
There are further improvements possible based on the scenario in #362 as a big problem with the implementation when setting multiple entities is that the number of checks via `GetEntry` grows with every extra entity.

That is to say, you add 10 items to an empty entity container:
- The first will check 0 items
- The second will check 1 item
- The third will check 2 items
- etc

We can avoid a bunch of checks by not checking the items we're adding against themselves while we're adding them. So if someone were to add 500,000 items, we don't actually need to check them against each other.

That said, without doing the checks, it does mean it is possible for someone to intentionally double up the exact same reference entity in the database (eg. the enumerable has the same entity twice). This could be avoided by chucking the whole thing in a `HashSet` perhaps but that will cause a large amount of allocations, especially for the case with 500,000 entities being added.