@Turnerj 2023-10-01
I would like to work out a good/better solution for adding many entries at once without a large number of allocations while addressing the underlying performance problem.